+++ java/src/main/java/io/cucumber/prettyformatter/LineBuilder.java
@@ -5,6 +5,7 @@ import io.cucumber.prettyformatter.Theme.Element;
 
 import java.util.function.Consumer;
 
+import static io.cucumber.prettyformatter.PrettyReportData.VISUAL_STATUS_ICON_LENGTH;
 import static java.lang.System.lineSeparator;
 
 final class LineBuilder {
@@ -25,6 +26,12 @@ final class LineBuilder {
         return append(keywordElement, keyword + ":")
                 .append(" ")
                 .append(nameElement, name);
+    }
+
+    LineBuilder statusIcon(String statusIcon) {
+        this.unstyledLength += VISUAL_STATUS_ICON_LENGTH;
+        builder.append(statusIcon);
+        return this;
     }
 
     LineBuilder addPaddingUpTo(int index) {