+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add status icons and plain text theme ([#3](https://github.com/cucumber/pretty-formatter/pull/3))
+
 
 ## [0.2.0] - 2025-07-07
 ### Changed